WIP commit for 3.1.0
The IV used in an AES context should be changed after each encryption operation,
which is not the case of the 2.1.0 version: This can lead to security holes in communications.

The 3.1.0 will focus on get this security flaw patched, and since it will consists
of modifying the communication format between client and server there will be no backwards compatibility.